+++ AiCard/AiCard/Controllers/CustomerTabsManageController.cs
@@ -10,15 +10,50 @@ using AiCard.Models;
 
 namespace AiCard.Controllers
 {
+    [Authorize]
     public class CustomerTabsManageController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
-        // GET: CustomerTabsManage
-        public ActionResult Index()
+        //获取当前登录的用户信息
+        AccountData AccontData
         {
-            var customerTabs = db.CustomerTabs.Include(c => c.Group);
-            return View(customerTabs.ToList());
+            get
+            {
+                return this.GetAccountData();
+            }
+        }
+        public void Sidebar(string name = "客户标签")
+        {
+            ViewBag.Sidebar = name;
+        }
+        // GET: CustomerTabsManage
+        [Authorize(Roles = SysRole.CustomerTabManageRead + "," + SysRole.ECustomerTabManageRead)]
+        public ActionResult Index(string filter, int? page = 1)
+        {
+            Sidebar();
+            var query = from ct in db.CustomerTabs
+                        from ctg in db.CustomerTabGroups
+                        where ct.GroupID == ctg.ID
+                        select new CustomerTabViewModel
+                        {
+                            ID = ct.ID,
+                            Name = ct.Name,
+                            Sort = ct.Sort,
+                            GroupName = ctg.Name,
+                            EnterpriseID = ctg.EnterpriseID
+                        };
+            if (!string.IsNullOrWhiteSpace(filter))
+            {
+                query = query.Where(s => s.Name.Contains(filter));
+            }
+            //如果是企业用户则只查询该企业信息
+            if (AccontData.UserType == Enums.UserType.Enterprise)
+            {
+                query = query.Where(s => s.EnterpriseID == AccontData.EnterpriseID);
+            }
+            var paged = query.OrderByDescending(s => s.ID).ToPagedList(page.Value, 10); ;
+            return View(paged);
         }
 
         // GET: CustomerTabsManage/Details/5
@@ -37,6 +72,7 @@ namespace AiCard.Controllers
         }
 
         // GET: CustomerTabsManage/Create
+        [Authorize(Roles = SysRole.CustomerTabManageCreate + "," + SysRole.ECustomerTabManageCreate)]
         public ActionResult Create()
         {
             ViewBag.GroupID = new SelectList(db.CustomerTabGroups, "ID", "Name");
@@ -48,8 +84,17 @@ namespace AiCard.Controllers
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID,GroupID,Name,Sort")] CustomerTab customerTab)
+        [Authorize(Roles = SysRole.CustomerTabManageCreate + "," + SysRole.ECustomerTabManageCreate)]
+        public ActionResult Create(CustomerTab customerTab)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if ((AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID) || AccontData.EnterpriseID <= 0)
+            {
+                return this.ToError("错误", "没有该操作权限", Url.Action("Index"));
+            }
+            Sidebar();
             if (ModelState.IsValid)
             {
                 db.CustomerTabs.Add(customerTab);
@@ -62,8 +107,17 @@ namespace AiCard.Controllers
         }
 
         // GET: CustomerTabsManage/Edit/5
+        [Authorize(Roles = SysRole.CustomerTabManageEdit + "," + SysRole.ECustomerTabManageEdit)]
         public ActionResult Edit(int? id)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if ((AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID) || AccontData.EnterpriseID <= 0)
+            {
+                return this.ToError("错误", "没有该操作权限", Url.Action("Index"));
+            }
+            Sidebar();
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -82,11 +136,29 @@ namespace AiCard.Controllers
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,GroupID,Name,Sort")] CustomerTab customerTab)
+        [Authorize(Roles = SysRole.CustomerTabManageEdit + "," + SysRole.ECustomerTabManageEdit)]
+        public ActionResult Edit(CustomerTab customerTab)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if ((AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID) || AccontData.EnterpriseID <= 0)
+            {
+                return this.ToError("错误", "没有该操作权限", Url.Action("Index"));
+            }
+            Sidebar();
             if (ModelState.IsValid)
             {
                 db.Entry(customerTab).State = EntityState.Modified;
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            var temp = db.CustomerTabs.FirstOrDefault(s => s.ID == customerTab.ID);
+            if (ModelState.IsValid)
+            {
+                temp.Name = customerTab.Name;
+                temp.Sort = customerTab.Sort;
+                temp.GroupID = customerTab.GroupID;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
@@ -95,25 +167,54 @@ namespace AiCard.Controllers
         }
 
         // GET: CustomerTabsManage/Delete/5
+        [Authorize(Roles = SysRole.CustomerTabMangeDelete + "," + SysRole.ECustomerTabMangeDelete)]
         public ActionResult Delete(int? id)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if ((AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID) || AccontData.EnterpriseID <= 0)
+            {
+                return this.ToError("错误", "没有该操作权限", Url.Action("Index"));
+            }
+            Sidebar();
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            CustomerTab customerTab = db.CustomerTabs.Find(id);
-            if (customerTab == null)
+            var query = (from ct in db.CustomerTabs
+                        from ctg in db.CustomerTabGroups
+                        where ct.GroupID == ctg.ID
+                        select new CustomerTabViewModel
+                        {
+                            ID = ct.ID,
+                            Name = ct.Name,
+                            Sort = ct.Sort,
+                            GroupName = ctg.Name,
+                            EnterpriseID = ctg.EnterpriseID
+                        }).FirstOrDefault();
+            //CustomerTab customerTab = db.CustomerTabs.Find(id);
+            if (query == null)
             {
                 return HttpNotFound();
             }
-            return View(customerTab);
+            return View(query);
         }
 
         // POST: CustomerTabsManage/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = SysRole.CustomerTabMangeDelete + "," + SysRole.ECustomerTabMangeDelete)]
         public ActionResult DeleteConfirmed(int id)
         {
+            var tempuser = db.Users.FirstOrDefault(s => s.Id == AccontData.UserID);
+            //防止企业用户串号修改
+            if ((AccontData.UserType == Enums.UserType.Enterprise
+                && tempuser.EnterpriseID != AccontData.EnterpriseID) || AccontData.EnterpriseID <= 0)
+            {
+                return this.ToError("错误", "没有该操作权限", Url.Action("Index"));
+            }
+            Sidebar();
             CustomerTab customerTab = db.CustomerTabs.Find(id);
             db.CustomerTabs.Remove(customerTab);
             db.SaveChanges();
